# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Visitor Management System API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/v1/department/{department_id}/users": {
      "get": {
        "summary": "Get all users for a specific department",
        "tags": [
          "Departments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "department_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Department ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of users for the department"
          }
        }
      }
    },
...
    "/api/v1/user": {
      "get": {
        "summary": "Get all users",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "List of users"
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          }
        }
      }
    },
...
    "/api/v1/user/{user_id}": {
      "get": {
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details"
          }
        }
      },
      "put": {
        "summary": "Update a user",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a user",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          }
        }
      }
    },
    "/api/v1/userTypes": {
      "get": {
        "summary": "Get all user types",
        "tags": [
          "UserTypes"
        ],
        "responses": {
          "200": {
            "description": "List of user types"
          }
        }
      },
      "post": {
        "summary": "Create a new user type",
        "tags": [
          "UserTypes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User type created"
          }
        }
      }
    },
    "/api/v1/userTypes/{user_type_id}": {
      "get": {
        "summary": "Get user type by ID",
        "tags": [
          "UserTypes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_type_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User type details"
          }
        }
      },
      "put": {
        "summary": "Update a user type",
        "tags": [
          "UserTypes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_type_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User type updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a user type",
        "tags": [
          "UserTypes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_type_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User type deleted"
          }
        }
      }
    },
...
  },
...
}